requirements-docling-full.txt
# requirements-docling-full.txt
# Docling Full Layer: Complete Docling with all features
# ~20 packages, ~1.2GB

# Base dependencies
pydantic>=2.12.0
orjson>=3.9.0
numpy>=1.26.0,<2.0.0
pandas>=2.2.0

# Complete Docling
docling>=2.25.0

# Full ML Stack
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.36.0
tensorflow>=2.14.0

# OCR and Vision
easyocr>=1.7.0
opencv-python>=4.8.0
pillow>=10.0.0
pytesseract>=0.3.10

# PDF Processing
pypdfium2>=4.0.0
pdfplumber>=0.10.0
pymupdf>=1.23.0

# Document processing
python-docx>=0.8.11
python-pptx>=0.6.21
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Text processing
nltk>=3.8.0
spacy>=3.7.0
